refactor: rename database model files and add new models file
- Rename GameSession.py to game_session.py
- Rename stationsTakeOvers.py to stations_take_overs.py
- Add new models.py file for shared database models
